refactor(cli): extract sentinel cluster + hoist buildHealthCtx (5.0-alpha)
Sixth cut of the 5.0-alpha cli.ts god-module split.

- Hoists buildHealthCtx (co-owned by cmdHealth, which stays in cli.ts) into the
  neutral cli-shared.ts, imported back by cmdHealth and by the new module — no
  commands -> cli.ts cycle.
- New src/commands/sentinel.ts: cmdSentinel (exported) + its run/install/status
  sub-handlers (module-private, routed via process.argv[3]).
- cli.ts imports cmdSentinel + buildHealthCtx back; COMMANDS table unchanged;
  prunes now-unused imports (SentinelSummary, HealthCtx type, writeFileSync).
- Dynamic imports re-pathed ./ -> ../ (sentinel.js, health.js, execFileNoThrow).

cli.ts: 3966 -> 3789 lines (-177; -2541 cumulative across PRs 1-6, from 6330).
Typecheck, eslint (0 errors), full suite (2580), prettier all green; kit
sentinel status + kit health smoke-tested.
